pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["maksimfadeev <7522688@gmail.com>"]
packages = [
    { include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
coverage = "^6.3.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

